@@ -1,5 +1,5 @@
-module github.com/notnil/chess
+module github.com/othomann/go-chess
 
-go 1.14
+go 1.20
 
 require github.com/ajstarks/svgo v0.0.0-20200320125537-f189e35d30ca
