@@ -7,7 +7,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/notnil/chess"
+	"github.com/othomann/go-chess"
 )
 
 // Cmd is a UCI compliant command
@@ -125,11 +125,12 @@ var (
 // The substrings "value" and "name" should be avoided in  and  to allow unambiguous parsing,
 // for example do not use  = "draw value".
 // Here are some strings for the example below:
-//    "setoption name Nullmove value true\n"
-//   "setoption name Selectivity value 3\n"
-//    "setoption name Style value Risky\n"
-//    "setoption name Clear Hash\n"
-//    "setoption name NalimovPath value c:\chess\tb\4;c:\chess\tb\5\n"
+//
+//	 "setoption name Nullmove value true\n"
+//	"setoption name Selectivity value 3\n"
+//	 "setoption name Style value Risky\n"
+//	 "setoption name Clear Hash\n"
+//	 "setoption name NalimovPath value c:\chess\tb\4;c:\chess\tb\5\n"
 type CmdSetOption struct {
 	Name  string
 	Value string
@@ -179,42 +180,42 @@ func (CmdPosition) ProcessResponse(e *Engine) error {
 // start calculating on the current position set up with the "position" command.
 // There are a number of commands that can follow this command, all will be sent in the same string.
 // If one command is not send its value should be interpreted as it would not influence the search.
-// * searchmoves  ....
-// 	restrict search to this moves only
-// 	Example: After "position startpos" and "go infinite searchmoves e2e4 d2d4"
-// 	the engine should only search the two moves e2e4 and d2d4 in the initial position.
-// * ponder
-// 	start searching in pondering mode.
-// 	Do not exit the search in ponder mode, even if it's mate!
-// 	This means that the last move sent in in the position string is the ponder move.
-// 	The engine can do what it wants to do, but after a "ponderhit" command
-// 	it should execute the suggested move to ponder on. This means that the ponder move sent by
-// 	the GUI can be interpreted as a recommendation about which move to ponder. However, if the
-// 	engine decides to ponder on a different move, it should not display any mainlines as they are
-// 	likely to be misinterpreted by the GUI because the GUI expects the engine to ponder
-//    on the suggested move.
-// * wtime
-// 	white has x msec left on the clock
-// * btime
-// 	black has x msec left on the clock
-// * winc
-// 	white increment per move in mseconds if x > 0
-// * binc
-// 	black increment per move in mseconds if x > 0
-// * movestogo
-//   there are x moves to the next time control,
-// 	this will only be sent if x > 0,
-// 	if you don't get this and get the wtime and btime it's sudden death
-// * depth
-// 	search x plies only.
-// * nodes
-//    search x nodes only,
-// * mate
-// 	search for a mate in x moves
-// * movetime
-// 	search exactly x mseconds
-// * infinite
-// 	search until the "stop" command. Do not exit the search without being told so in this mode!
+//   - searchmoves  ....
+//     restrict search to this moves only
+//     Example: After "position startpos" and "go infinite searchmoves e2e4 d2d4"
+//     the engine should only search the two moves e2e4 and d2d4 in the initial position.
+//   - ponder
+//     start searching in pondering mode.
+//     Do not exit the search in ponder mode, even if it's mate!
+//     This means that the last move sent in in the position string is the ponder move.
+//     The engine can do what it wants to do, but after a "ponderhit" command
+//     it should execute the suggested move to ponder on. This means that the ponder move sent by
+//     the GUI can be interpreted as a recommendation about which move to ponder. However, if the
+//     engine decides to ponder on a different move, it should not display any mainlines as they are
+//     likely to be misinterpreted by the GUI because the GUI expects the engine to ponder
+//     on the suggested move.
+//   - wtime
+//     white has x msec left on the clock
+//   - btime
+//     black has x msec left on the clock
+//   - winc
+//     white increment per move in mseconds if x > 0
+//   - binc
+//     black increment per move in mseconds if x > 0
+//   - movestogo
+//     there are x moves to the next time control,
+//     this will only be sent if x > 0,
+//     if you don't get this and get the wtime and btime it's sudden death
+//   - depth
+//     search x plies only.
+//   - nodes
+//     search x nodes only,
+//   - mate
+//     search for a mate in x moves
+//   - movetime
+//     search exactly x mseconds
+//   - infinite
+//     search until the "stop" command. Do not exit the search without being told so in this mode!
 type CmdGo struct {
 	SearchMoves    []*chess.Move
 	Ponder         bool
@@ -312,7 +313,7 @@ func (CmdGo) ProcessResponse(e *Engine) error {
 }
 
 func parseIDLine(s string) (string, string, error) {
-	if strings.HasPrefix(s, "id") == false {
+	if !strings.HasPrefix(s, "id") {
 		return "", "", errors.New("uci: invalid id line")
 	}
 	parts := strings.Split(s, " ")
